Refactord to using cogs
