--- tamago/lib/__init__.py
@@ -1,5 +1,0 @@
-"""
-init - version info
-"""
-VERSION_INFO = (0, 0, 1)
-VERSION = '.'.join(str(c) for c in VERSION_INFO)
