--- tamago/lib/plugins/music_bot.py
@@ -1,9 +1,0 @@
-import discord
-from discord.ext import commands
-
-class Music:
-    def __init__(self, client):
-        self.client = client
-
-
-
